refactor: Mock 파일 이동 및 중복 정리 #139
